requirements.txt
daphne
django-rest-framework
django-cors-headers
django-bootstrap-v5
django-channels
django-snippet
django-environ
django-filter
markdown
psycopg
psycopg2-binary
Pillow